Drop the native API emulation
It's not possible to perfectly emulate live collections and even without
that, this is never going to work on older browsers. The plan going
forward is to only have the jQuery plugin and to use browser
implementations (where available) to make that faster, similar to how
jQuery will use querySelector iff available.