Revamp the property crawling algorithm to not remove loops
After the implementation feedback in
<http://lists.whatwg.org/pipermail/whatwg-whatwg.org/2011-June/032257.html>
we're abandoning the idea of doing loop removal as part of the
.properties API and can simplify the crawling a fair bit. The algorithm
implemented here is black-box-testably the same as what Opera has
implemented and what I intend to suggest for the spec.

This will break at least JSON and RDF export, following commits will
explore how to deal with loops in those algorithms.